Turn on bluetooth when required after successful onboarding

See merge request convexityteam/chats-field-app!24

+++ app/src/main/java/chats/cash/chats_field/views/auth/viewmodel/RegisterViewModel.kt
@@ -96,7 +96,7 @@ class RegisterViewModel(
                     campaign,
                     pin
                 )
-                data.apply {
+                if (data is ApiResponse.Success<RegisterResponse>) {
                     profile_pic.delete()
                 }
                 Timber.d("data: $data")
